regroup files
